+++ packages/tiptap-extensions/web/src/bulleted-list.ts
@@ -1,9 +1,10 @@
-import { wrappingInputRule } from '@tiptap/core';
+import { InputRule } from '@tiptap/core';
 import type { RicosExtension } from 'ricos-tiptap-types';
 import bulletedListDataDefaults from 'ricos-schema/dist/statics/bulleted_list.defaults.json';
 import type { DOMOutputSpec } from 'prosemirror-model';
 import { Node_Type } from 'ricos-schema';
 import styles from './statics/styles.scss';
+import { createListInputRuleHandler } from './list-input-rule-handler';
 
 export interface BulletListOptions {
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
@@ -77,9 +78,9 @@ export const bulletedList: RicosExtension = {
 
       addInputRules() {
         return [
-          wrappingInputRule({
+          new InputRule({
             find: inputRegex,
-            type: this.type,
+            handler: createListInputRuleHandler({ type: this.type }),
           }),
         ];
       },